+++ modules/contracts/src.sol/CMCCore.sol
@@ -6,7 +6,7 @@ import "./interfaces/ICMCCore.sol";
 import "./ReentrancyGuard.sol";
 
 contract CMCCore is ReentrancyGuard, ICMCCore {
-  address constant invalidParticipant = address(1);
+  address private immutable mastercopyAddress;
 
   address internal alice;
   address internal bob;
@@ -15,12 +15,12 @@ contract CMCCore is ReentrancyGuard, ICMCCore {
   ///         Nonzero address also prevents the mastercopy from being setup
   ///         Only setting alice is sufficient, setting bob too wouldn't change anything
   constructor () {
-      alice = invalidParticipant;
+      mastercopyAddress = address(this);
   }
 
   // Prevents us from calling methods directly from the mastercopy contract
   modifier onlyViaProxy {
-    require(alice != address(1), "Mastercopy: ONLY_VIA_PROXY");
+    require(address(this) != mastercopyAddress, "Mastercopy: ONLY_VIA_PROXY");
     _;
   }
 
@@ -28,10 +28,10 @@ contract CMCCore is ReentrancyGuard, ICMCCore {
   /// @param _alice: Address representing user with function deposit
   /// @param _bob: Address representing user with multisig deposit
   function setup(address _alice, address _bob) external override onlyViaProxy {
-    ReentrancyGuard.setup();
     require(alice == address(0), "CMCCore: ALREADY_SETUP");
     require(_alice != address(0) && _bob != address(0), "CMCCore: INVALID_PARTICIPANT");
     require(_alice != _bob, "CMCCore: IDENTICAL_PARTICIPANTS");
+    ReentrancyGuard.setup();
     alice = _alice;
     bob = _bob;
   }