+++ modules/contracts/src.sol/interfaces/ICMCAccountant.sol
@@ -1,9 +1,0 @@
-// SPDX-License-Identifier: UNLICENSED
-pragma solidity ^0.7.1;
-pragma experimental ABIEncoderV2;
-
-import "./IAssetTransfer.sol";
-import "./ICMCDeposit.sol";
-import "./ICMCWithdraw.sol";
-
-interface ICMCAccountant is IAssetTransfer, ICMCDeposit, ICMCWithdraw {}