+++ modules/contracts/src.sol/testing/TestChannel.sol
@@ -10,9 +10,9 @@ import "../interfaces/ITestChannel.sol";
 /// @notice This contract will help test the `ChannelMastercopy` contract and
 ///         the associated bits of functionality. This contract should *only*
 ///         contain aliases to internal functions that should be unit-tested,
-///         like the `makeEmergencyWithdrawable` call on `AssetTransfer.sol`. Using this
+///         like the `makeEmergencyWithdrawable` call on `CMCAsset.sol`. Using this
 ///         contract will help reduce the amount of boilerplate needed to test
-///         component functionality. For example, `AssetTransfer.sol` is only
+///         component functionality. For example, `CMCAsset.sol` is only
 ///         able to be tested via the adjudicator in many practical cases.
 ///         Creating a helper function allows for easier testing of only
 ///         that functionality.