+++ modules/documentation/docs/changelog.md
@@ -3,6 +3,10 @@
 ## Next Release
 
 - use individual ethers packages instead of ethers
+- various smaller contract changes
+- simplified inheritance hierarchy for channel mastercopy
+- cheaper `onlyViaProxy` modifier
+- saturated addition for emergency-withdrawable amounts
 
 ## 0.0.13